fixed error in UseEngineResponse return value

+++ src/main/java/it/polimi/ingsw/galaxytruckerproject/network/SOCKET/message/UseEngineResponse.java
@@ -19,7 +19,7 @@ public class UseEngineResponse extends Message{
         return playerName;
     }
 
-    public float getNumEngine() {
+    public int getNumEngine() {
         return numEngine;
     }
